requirements.txt
nltk==3.7
pyenchant==3.2.2
PyQt5==5.15.7
